lots of problems fixed... also should've commit more
