+++ src/builtins/exit.c
@@ -6,7 +6,7 @@
 /*   By: gakarbou <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/24 15:05:06 by gakarbou          #+#    #+#             */
-/*   Updated: 2025/02/27 18:58:30 by gakarbou         ###   ########.fr       */
+/*   Updated: 2025/02/27 19:05:29 by gakarbou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
