+++ src/minishell.c
@@ -6,7 +6,7 @@
 /*   By: lroussel <lroussel@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/22 22:50:10 by lroussel          #+#    #+#             */
-/*   Updated: 2025/02/27 14:30:50 by gakarbou         ###   ########.fr       */
+/*   Updated: 2025/02/27 20:13:30 by gakarbou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,7 +35,8 @@ void	launch(t_list *envp, t_main_envp *imp)
 			free(res);
 			continue ;
 		}
-		str = parse_commands(res, envp, imp);
+		str = execute_command(res, &envp, imp);
+		free(res);
 		if (str)
 			(printf("%s", str), free(str));
 	}
