--- src/parse_commands.c
@@ -1,50 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   parse_commands.c                                   :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: gakarbou <marvin@42.fr>                    +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/02/23 12:55:20 by gakarbou          #+#    #+#             */
-/*   Updated: 2025/02/27 18:51:02 by gakarbou         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "minishell.h"
-
-char	*handle_builtins(int code, char **argv, t_list **envp, t_main_envp *imp)
-{
-	if (code == 1)
-		return (ms_echo(argv));
-	else if (code == 2)
-		return (ms_cd(argv));
-	else if (code == 3)
-		return (ms_pwd());
-	else if (code == 4)
-		return (ms_export(argv, envp));
-	else if (code == 5)
-		return (ms_unset(argv, envp, imp));
-	else if (code == 6)
-		return (ms_env(argv, *envp));
-	else if (code == 7)
-		return (ms_exit(argv));
-	return (NULL);
-}
-
-char	*parse_commands(char *str, t_list *envp, t_main_envp *imp)
-{
-	t_int_tab	tabe;
-	char		**argv;
-	char		*dest;
-
-	tabe = init_int_tab();
-	argv = create_command_argv(str, envp, imp);
-	tabe.res = check_built_in(&argv[0]);
-	if (tabe.res)
-		dest = handle_builtins(tabe.res, argv, &envp, imp);
-	else
-		dest = execute_command(argv, imp);
-	while (argv[++tabe.i])
-		free(argv[tabe.i]);
-	return (free(argv), dest);
-}
