+++ src/var_parser.c
@@ -6,43 +6,37 @@
 /*   By: gakarbou <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/23 03:46:01 by gakarbou          #+#    #+#             */
-/*   Updated: 2025/02/25 17:42:13 by gakarbou         ###   ########.fr       */
+/*   Updated: 2025/02/27 20:52:05 by gakarbou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-char	*parse_command(char *command, t_list *envp, t_main_envp *imp)
+static char	*handle_var_commands(char *command, t_list **envp, t_main_envp *imp)
 {
-	char	**argv;
 	char	*str;
-	int		len;
 
-	len = ft_strlen(command);
-	command[len - 1] = 0;
-	argv = create_command_argv(command, envp, imp);
-	str = execute_command(argv, imp);
-	len = -1;
-	while (argv[++len])
-		free(argv[len]);
-	free(argv);
+	command[ft_strlen(command) - 1] = 0;
+	str = execute_command(command, envp, imp);
 	return (str);
 }
 
-char	*parse_var(char *var_name, t_list *envp, t_main_envp *imp)
+char	*parse_var(char *var_name, t_list **envp, t_main_envp *imp)
 {
+	t_list	*cur;
 	char	*cur_env;
 	int		len;
 
 	if (var_name[0] == '(')
-		return (parse_command(var_name + 1, envp, imp));
+		return (handle_var_commands(var_name + 1, envp, imp));
 	len = ft_strlen(var_name);
-	while (envp)
+	cur = *envp;
+	while (cur)
 	{
-		cur_env = (char *)envp->content;
+		cur_env = (char *)cur->content;
 		if ((cur_env[len] == '=') && (!ft_strncmp(cur_env, var_name, len)))
 			return (ft_strdup(cur_env + len + 1));
-		envp = envp->next;
+		cur = cur->next;
 	}
 	return (NULL);
 }
